requirements/default.txt
-r base.txt

# These dependencies are explicitly included in code.

argparse==1.2.1 	# Python Software Foundation License
boto3==1.4.4 # Apache 2.0
ciso8601==1.0.3         # MIT
edx-opaque-keys==0.4    # AGPL
edx-ccx-keys==0.2.1     # AGPL
elasticsearch==1.7.0    # Apache
filechunkio==1.8	# MIT
graphitesend==0.10.0 # Apache
html5lib==1.0b3 	# MIT
isoweek==1.3.1		# BSD
http://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-1.2.2.zip  	# GPL v2 with FOSS License Exception
numpy==1.11.3 		# BSD
pandas==0.13.0 		# BSD
paypalrestsdk==1.9.0    # Paypal SDK License
pygeoip==0.3.2 		# LGPL
python-cjson==1.1.0	# LGPL
python-dateutil==2.6.0 	# BSD
python-gnupg==0.3.9	# BSD
pytz==2016.10		# ZPL
requests==2.12.4        # Apache 2.0
sailthru-client==2.3.1
six==1.10.0		# MIT
stevedore==1.19.1 	# Apache 2.0
ua-parser==0.3.6 	# Apache
urllib3==1.19.1         # MIT
user-agents==0.3.2	# MIT
vertica-python==0.6.11  # MIT
yarn-api-client==0.2.3 # BSD

git+https://github.com/edx/luigi.git@a73700ca51685974220ef6069d2f078312055444#egg=luigi 		# Apache License 2.0
git+https://github.com/edx/pyinstrument.git@a35ff76df4c3d5ff9a2876d859303e33d895e78f#egg=pyinstrument     # BSD

# These dependencies are pulled in by the above requirements,
# and are pinned here.

distlib==0.2.4
future==0.16.0
pymongo==3.4.0		# Apache 2.0
pyOpenSSL==16.2.0       # Apache 2.0
pbr==1.10.0 		# Apache

# These dependencies are already pulled in by earlier requirements (e.g.
# and are NOT pinned here.

# cffi==1.9.1             # MIT
# cryptography==1.7.1     # BSD or Apache 2.0
# enum34==1.1.6           # BSD
# idna==2.2               # BSD-like
# pip==9.0.1
# pyasn1==0.1.9           # BSD
# pycparser==2.17         # BSD
# setuptools==32.3.1

# These dependencies are not used?

# backports_abc==0.5
# certifi==2016.9.26
# mechanize==0.2.5	# BSD
# ndg-httpsclient==0.4.2  # BSD
# singledispatch==3.4.0.3
# tornado==4.4.2 		# Apache 2.0
# wsgiref==0.1.2 		# ZPL
